Only show extension points that has a template.
